Heavy refactoring. Still not there yet...
