+++ src/Orleans.Runtime/Utilities/ChangeFeedEntry.cs
@@ -1,0 +1,13 @@
+using System.Threading.Tasks;
+
+namespace Orleans.Runtime
+{
+    public abstract class ChangeFeedEntry<T>
+    {
+        public abstract bool HasValue { get; }
+
+        public abstract T Value { get; }
+
+        public abstract Task<ChangeFeedEntry<T>> NextAsync();
+    }
+}
