+++ test/TesterInternal/GeoClusterTests/GatewaySelectionTest.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.Net;
 using Orleans.Runtime;
 using Orleans.Runtime.MembershipService;
+using Orleans.Runtime.MultiClusterNetwork;
 using Xunit;
-using UpdateFaultCombo = Orleans.Runtime.MembershipService.MembershipOracleData.UpdateFaultCombo;
 
 namespace UnitTests.GeoClusterTests
 {
@@ -30,13 +30,13 @@ namespace UnitTests.GeoClusterTests
                 SiloAddress.New(new IPEndPoint(3,234),1),
                 SiloAddress.New(new IPEndPoint(4,234),1),
             };
-            
-            Func<SiloAddress,UpdateFaultCombo> group = (SiloAddress a) => new UpdateFaultCombo(a.Endpoint.Port, a.Generation);
+
+            Func<SiloAddress, UpdateFaultCombo> group = (SiloAddress a) => new UpdateFaultCombo(a.Endpoint.Port, a.Generation);
 
             // randomize order
             var r = new Random();
-            var randomized = new SortedList<int,SiloAddress> ();
-            foreach(var c in candidates)
+            var randomized = new SortedList<int, SiloAddress>();
+            foreach (var c in candidates)
                 randomized.Add(r.Next(), c);
 
             var x = MembershipHelper.DeterministicBalancedChoice(randomized.Values, 10, group);
